backend/requirements.txt
django===5.2.6
djangorestframework===3.16.1
django-allauth===65.11.2
djangorestframework-simplejwt===5.5.1
django-anymail===13.1
pillow===11.3.0
cloudinary===1.44.1
django-cloudinary-storage===0.3.0
django-filter===25.1
requests===2.32.5
django-cors-headers===4.9.0
drf-yasg===1.21.11